Payment Integration

+++ WorkSpaceWebAPI/Models/Review.cs
@@ -4,11 +4,11 @@ namespace WorkSpaceWebAPI.Models
 {
     public class Review
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         [ForeignKey("User")]
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         [ForeignKey("Room")]
-        public long RoomId { get; set; }
+        public int RoomId { get; set; }
         public int Rating { get; set; }
         public string? Comment { get; set; }
         public DateTime CreatedAt { get; set; }
